Move getNotConfiguredTimeline to TileViews.kt.
